Changes the way the export line was written.
When building for ES modules, we append `.js` to all imports as part of the build process.  However, we do not append `.js` to exports.  We could fix the build such that it appends `.js` to exports, but it is simpler for now to just change the export since the previous way was probably wrong anyway (did not follow DRY).