Fixes incompatibilities between Parity and Geth and a bug in nonce selection strategy.
Parity will return a `Partial<TransactionReceipt>` if you call `eth_getTransactionReceipt` prior to the transaction being mined.  Geth on the other hand will return `null`.  This library will now treat a `Partial<TransactionReceipt>` as `null` so that both Geth and Parity exhibit the same behavior from the perspective of library users.

When we call `eth_getTransactionCount` in order to acquire the nonce for a new transaction, we want to include pending transactions (and thus the pending block) so that if a user fires multiple transactions back to back we'll get a new nonce for each transaction instead of the same nonce repeated until a transaction is mined.