Changes the expected signature to contain parity, not v.
This change makes it so this library calculates `v` on behalf of the user, rather than expecting the signer to do it.  This makes it simpler to implement a signer, and doesn't rely on the signer correctly applying EIP-155.  It also means you can use any secp256k1 signer rather than being limited to ethereum aware signers and it also means the signer doesn't need to extract the chainId from the RLP encoded transaction being signed.