Makes some base RPC methods use the signer if available.
`coinbase`, `getAccounts`, `sendTransaction`, and `sign` all now use the signer if available rather than just falling through to the backing JSON-RPC server.

Also updates ethereum-types and fixes a bug in `.coinbase` causing it to now return `null` instead of throwing an error if `coinbase` is not available from the backing server.