Adds gas estimate provider and reafctors constructor.
With the addition of yet another optional member to the constructor, I decided it was time to refactor it into an options object for ease of use.

Rather than just running estimate gas naively, we give the user an opportunity to do their own estimation.  We give them the original estimator function so they can do normal estimation with some modifier, or they could inspect the transaction and manually set gas costs based on the transaction being executed.