Changes `id` in all requests to be `0` instead of `null` to deal with non-compliant servers.
Some JSON-RPC servers are not spec compliant and they disallow `null` for the `id`.  In order to be maximally compliant, we are using `0` for the ID instead.  It is worth noting that the JSON-RPC specification does discourage the use of `null`, even though it explicitly states it is allowed so this change also brings us in line with the spirit of the specification, so it isn't a bad thing really.