Adds handling of Parity contract revert errors.
Parity, unlike Geth/Nethermind, will return a JSON-RPC error with a data property of `Reverted <encoded Error('message')>` when a contract reverts with a message.  If we don't handle this in this library, then it will get swallowed and the caller will not be able to handle it.  So, we handle it here and include the decoded revert message in the thrown error.