Works around a bug in Nethermind while waiting for a fix.
Pulls in the change from https://github.com/Zoltu/ethereum-types/commit/a88c1739c2d4b66e6a87b56fcc5c480e4e518e67.

Nethermind returns `null` instead of `0` when asking for a proof of a merkle node that is unset.  This change makes it so we default to `0` in that situation.